Add plan ownership system

+++ dsb/modules/stable/planner.py
@@ -6,7 +6,7 @@ from telegram import Update
 from telegram.ext import ContextTypes
 from dsb.types.lesson import Lesson
 from dsb.types.plan import Plan
-from dsb.types.module import BaseModule, prevent_edited
+from dsb.types.module import BaseModule, prevent_edited, admin_only
 if TYPE_CHECKING:
     from dsb.dsb import DSB
 
@@ -52,9 +52,9 @@ class Planner(BaseModule):
             "transfer_plan": "Transfer a plan to another group"
         }
 
-    def create_plan(self, name: str, group_id: int) -> bool:
+    def create_plan(self, name: str, group_id: int, user_id: int) -> bool:
         """ Create a new lesson plan """
-        new_plan = Plan(name)
+        new_plan = Plan(name, owner=user_id)
         return self._db.save(new_plan, f"{group_id}/plans", name)
 
     def get_plan(self, name: str, group_id: int) -> Plan | None:
@@ -130,7 +130,7 @@ class Planner(BaseModule):
             await update.message.reply_text("A plan with that name already exists")
             return
 
-        self.create_plan(plan_name, new_group)
+        self.create_plan(plan_name, new_group, update.effective_user.id)
         self.update_plan(plan_name, new_group, plan)
         await update.message.set_reaction("👍")
 
@@ -157,7 +157,7 @@ class Planner(BaseModule):
             await update.message.reply_text("A plan with that name already exists")
             return
 
-        if self.create_plan(plan_name, group_id):
+        if self.create_plan(plan_name, group_id, update.effective_user.id):
             await update.message.set_reaction("👍")
         else:
             await update.message.reply_text("An error occurred")
@@ -176,6 +176,17 @@ class Planner(BaseModule):
             plan_name = " ".join(args)
 
         group_id = update.effective_chat.id
+        plan = self.get_plan(plan_name, group_id)
+
+        if plan is None:
+            await update.message.reply_text(f"There is no plan with name {plan_name}")
+            return
+
+        user_id = update.effective_chat.id
+        if user_id != plan.owner and user_id not in self.config["admins"]:
+            await update.message.reply_text("This plan does not belong to you")
+            return
+
         if self.delete_plan(plan_name, group_id):
             await update.message.set_reaction("👍")
         else:
@@ -201,12 +212,17 @@ class Planner(BaseModule):
             await update.message.reply_text("Plan not found")
             return
 
+        user_id = update.effective_user.id
+        if plan.owner != user_id and user_id not in self.config["admins"]:
+            await update.message.reply_text("This is not your plan")
+
         if self.get_plan(kwargs.get("new_name"), group_id):
             await update.message.reply_text("A plan with that name already exists")
 
         self.update_plan(plan_name, group_id, plan, kwargs.get("new_name"))
         await update.message.set_reaction("👍")
 
+    @admin_only
     @prevent_edited
     async def _delete_all(self, update: Update, _) -> None:
         """ Delete all lesson plans """
@@ -343,6 +359,11 @@ class Planner(BaseModule):
             await update.message.reply_text(f"Plan {plan_name} not found")
             return
 
+        user_id = update.effective_user.id
+        if plan.owner != user_id and not user_id in self.config["admins"]:
+            await update.message.reply_text("This plan does not belong to you")
+            return
+
         args, kwargs = self._get_args(context)
         if "idx" in kwargs:
             idx = int(kwargs.get("idx"))
@@ -399,6 +420,11 @@ class Planner(BaseModule):
 
         if not plan:
             await update.message.reply_text(f"Plan {plan_name} not found")
+            return
+
+        user_id = update.effective_user.id
+        if plan.owner != user_id and not user_id in self.config["admins"]:
+            await update.message.reply_text("This plan does not belong to you")
             return
 
         args, kwargs = self._get_args(context)
@@ -465,6 +491,11 @@ class Planner(BaseModule):
 
         if not plan:
             await update.message.reply_text(f"Plan {plan_name} not found")
+            return
+
+        user_id = update.effective_user.id
+        if plan.owner != user_id and not user_id in self.config["admins"]:
+            await update.message.reply_text("This plan does not belong to you")
             return
 
         args, kwargs = self._get_args(context)
